refactor: complete handler reorganization and cleanup (#709)
Phase 3: Handler Reorganization
- Move handler pipeline from pkg/handlers/pipeline/ to pkg/execution/pipeline/
  - execute.go, execute_test.go, and integration.go moved
  - Updated package documentation
- Update imports in pkg/packs/commands/on.go
- Remove old pkg/handlers/pipeline/ directory

Phase 4: Cleanup and Documentation
- Run go mod tidy to clean dependencies
- Create refactoring_summary.md documenting all changes
- Verify all packages build successfully

This completes the major refactoring effort to improve code organization:
- Phase 1: Extracted business logic from types
- Phase 2: Reorganized pack structure and eliminated core package
- Phase 3: Moved handler pipeline to execution package
- Phase 4: Cleaned up and documented changes

Known issue: dispatcher couldn't be moved to execution package due to
import cycles. This would require additional architectural changes.
