+++ refactoring_summary.md
@@ -1,0 +1,98 @@
+# Dodot Refactoring Summary
+
+## Overview
+This document summarizes the major refactoring completed to improve the architecture and organization of the dodot codebase.
+
+## Changes Made
+
+### Phase 1: Type Extraction (Completed)
+1. **Business Logic Extraction**
+   - Moved business logic from data types to dedicated manager/aggregator classes
+   - Created `pkg/execution/context/manager.go` for ExecutionContext operations
+   - Created `pkg/execution/results/aggregator.go` for PackExecutionResult operations
+
+2. **Rules Package Creation**
+   - Created `pkg/rules/` to separate rule matching from handler execution
+   - Moved scanner.go, config.go, and rule types from handlers/pipeline
+
+3. **Dependency Resolution**
+   - Moved HandlerCategory type from handlers to operations package
+   - Resolved circular dependencies between handlers and operations
+
+### Phase 2: Pack Reorganization (Completed)
+1. **Pipeline → Execution Rename**
+   - Renamed `pkg/packs/pipeline/` to `pkg/packs/execution/`
+   - Updated all references and imports
+
+2. **Command Consolidation**
+   - Moved pipeline commands to main commands directory
+   - Created `pkg/packs/operations/` for pack operation logic
+
+3. **Core Package Elimination**
+   - Moved pack discovery to `pkg/packs/discovery/`
+   - Moved output generation to `pkg/ui/output/`
+   - Moved initialization to `pkg/handlers/`
+   - Successfully removed the core package
+
+### Phase 3: Handler Reorganization (Completed)
+1. **Handler Pipeline Movement**
+   - Moved handler pipeline from `pkg/handlers/pipeline/` to `pkg/execution/pipeline/`
+   - Updated all imports and references
+
+## Current Structure
+
+```
+pkg/
+├── dispatcher/          # Still in use (couldn't be moved due to import cycles)
+├── execution/          
+│   ├── context/        # Execution context management
+│   ├── pipeline/       # Handler execution pipeline
+│   ├── results/        # Result aggregation
+│   └── status.go       # Execution status types
+├── handlers/           
+│   ├── lib/           # Individual handler implementations
+│   ├── api.go         # Handler registry
+│   └── init.go        # Handler initialization
+├── operations/         # Operation definitions and execution
+├── packs/             
+│   ├── commands/      # Pack command implementations
+│   ├── discovery/     # Pack discovery logic
+│   ├── execution/     # Pack execution orchestration
+│   └── operations/    # Pack-specific operations
+├── rules/             # Rule matching system
+└── ui/
+    └── output/        # Output generation (config, snippets)
+```
+
+## Benefits Achieved
+
+1. **Clearer Separation of Concerns**
+   - Data types are pure data structures
+   - Business logic is in dedicated service classes
+   - Rule matching is separate from execution
+
+2. **Better Package Organization**
+   - Related functionality is grouped together
+   - Package names reflect their purpose
+   - Reduced coupling between packages
+
+3. **Improved Testability**
+   - Business logic can be tested independently
+   - Cleaner interfaces between components
+
+## Known Issues
+
+1. **Import Cycles**
+   - The dispatcher couldn't be moved to execution package due to circular dependencies
+   - This would require further architectural changes to resolve
+
+2. **Incomplete Implementations**
+   - Some stub implementations were added (e.g., in operations/status_impl.go)
+   - These need to be properly implemented
+
+## Next Steps
+
+1. Complete the stub implementations in the operations package
+2. Consider further refactoring to resolve the dispatcher import cycle
+3. Update all documentation to reflect the new structure
+4. Add architecture diagrams showing the new component relationships
